js: update api
