+++ packages/async-storage/android/build.gradle
@@ -1,6 +1,6 @@
 buildscript {
-    ext.getExtOrDefault = { name ->
-        return rootProject.ext.has(name) ? rootProject.ext.get(name) : "2.2.10"
+    ext.getExtOrDefault = {name ->
+        return rootProject.ext.has(name) ? rootProject.ext.get(name) : project.properties['AsyncStorage_' + name]
     }
 
     repositories {
@@ -21,8 +21,8 @@ apply plugin: "kotlin-android"
 
 apply plugin: "com.facebook.react"
 
-def getExtOrIntegerDefault(name, defaultValue) {
-    return rootProject.ext.has(name) ? rootProject.ext.get(name) : defaultValue
+def getExtOrIntegerDefault(name) {
+    return rootProject.ext.has(name) ? rootProject.ext.get(name) : (project.properties["AsyncStorage_" + name]).toInteger()
 }
 
 def isNewArchitectureEnabled() {
@@ -32,11 +32,11 @@ def isNewArchitectureEnabled() {
 android {
     namespace "org.asyncstorage"
 
-    compileSdkVersion getExtOrIntegerDefault("compileSdkVersion", 36)
+    compileSdkVersion getExtOrIntegerDefault("compileSdkVersion")
 
     defaultConfig {
-        minSdkVersion getExtOrIntegerDefault("minSdkVersion", 24)
-        targetSdkVersion getExtOrIntegerDefault("targetSdkVersion", 36)
+        minSdkVersion getExtOrIntegerDefault("minSdkVersion")
+        targetSdkVersion getExtOrIntegerDefault("targetSdkVersion")
     }
 
     buildFeatures {
@@ -84,9 +84,6 @@ repositories {
     google()
 }
 
-def kotlin_version = getExtOrDefault("kotlinVersion")
-
 dependencies {
     implementation "com.facebook.react:react-android"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
