+++ packages/async-storage/android/gradle.properties
@@ -1,0 +1,4 @@
+AsyncStorage_kotlinVersion=2.2.10
+AsyncStorage_minSdkVersion=24
+AsyncStorage_targetSdkVersion=36
+AsyncStorage_compileSdkVersion=36
