+++ packages/async-storage/src/AsyncStorage.ts
@@ -1,7 +1,7 @@
 /**
  * Asynchronous key-value storage interface compatible with the Web Storage API.
  */
-export interface PersistentAsyncStorage {
+export interface AsyncStorage {
   /**
    * Retrieves a single item from storage.
    * @param key - The key identifying the stored value.
